Bruno.jl Documentation

Missing docstring.

Missing docstring for Bruno.DataGeneration.ParamLogDiff. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Bruno.DataGeneration.MovingBlockBootstrap. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Bruno.DataGeneration.CircularBlockBootstrap. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Bruno.DataGeneration.StationaryBootstrap. Check Documenter's build log for details.

Bruno.DataGeneration.getData — Function

Description

getData is a function that generates data according to the parameter type

Syntax

data = getData(Param)
data = getData(Param, nSimulation)

Positional Inputs

Param::DataGenerator: Parameters that describe the desired data generating process

nSimulation::Integer: nSimulation is the number of simulations to run.

Possible DataGenInput parameter types are

::LogDiffInput - log-normal diffusion process

::BootstrapInput{MovingBlock}

::BootstrapInput{CircularBlock}

::BootstrapInput{Stationary}

DataGenInput parameter types can be constructed directly or with data_gen_input() function.

Outputs

data::AbstractArray: nTimeStep x nSimulation Real valued array, where each column contains the data for one simulation, and each row contains data for each timestep

Example

# initialize parameters
nTimeStep = 100
Param1 = LogDiffInput(nTimeStep)
Param2 = LogDiffInput(nTimeStep, volatility=0)

# create two datasets, one with default values, the second with no volatility
data1 = getData(Param1)
data2 = getData(Param2)

# create a third dataset with 100 simulation runs 
nSimulation = 100
data3 = getData(Param1,nSimulation)

# plot results
plt = plot(data3, show=true, color="blue", legend=false)
plot!(plt, data1, show=true, color="red", legend=false, linewidth=3)

Bruno.DataGeneration.getTime — Function

Description

getTime is a function that generates a corresponding date-time array for data generated with a ParamLogDiff input in getData

Syntax

data = getTime(Param)
data = getTime(Param, initial)

Positional Inputs

Param::LogDiffInput: Parameters that describe the desired log-diffusion process

initial::DateTime: The time that corresponds to the initial parameter in getData

Outputs

time::Array{DateTime,2}: nStepSize x 1 DateTime array, where each value is the time for each row of data returned by getData.

Example

import Plots.plot
import Plots.plot!
using Bruno.DataGeneration.LogDiffusion

# initialize parameters
nTimeStep = 100
Param1 = ParamLogDiff(nTimeStep)
Param2 = ParamLogDiff(nTimeStep, volatility=0)

# create two datasets, one with default values, the second with no volatility
data1 = getData(Param1)
data2 = getData(Param2)

# create a third dataset with 100 simulation runs 
nSimulation = 100
data3 = getData(Param1, nSimulation)

# get time axis for each dataset
time1 = getTime(Param1)
time2 = getTime(Param2)
time3 = getTime(Param1)

# plot results
plt = plot(time3, data3, show=true, color="blue", legend=false)
plot!(plt, time1, data1, show=true, color="red", legend=false, linewidth=3)

Documentation for Bruno.jl