Type System

Types

Base Assets (Widgets)

Bruno.Instruments.Widget — Type

Widgets are the root asset at the heart of the package. A 'Widget' can be any real world finicial object such as a stock, or commodity. For a list of possible subtypes use subtypes(Widget)

Bruno.Instruments.Stock — Type

Stock <: Widget

Widget subtype. Used as a base or root asset for FinancialInstrument

Bruno.Instruments.Commodity — Type

Commodity <: Widget

Widget subtype. Used as a base or root asset for FinancialInstrument

Bruno.Instruments.Bond — Type

Bond <: Widget

Widget subtype. Used as a base or root asset for FinancialInstrument

Financial Instruments (Derivatives)

Abstract types

Bruno.Instruments.FinancialInstrument — Type

FinancialInstrument is the supertype for any instrument that uses a base asset (widget) in its definition (like a financial derivative)

Bruno.Instruments.Option — Type

Option <: FinancialInstrument

abstract FinancialInstrument subtype. Supertype of all options contract types

Bruno.Instruments.CallOption — Type

CallOption{T <: Widget} <: Option

abstract Option subtype. Super type for all call options types

Bruno.Instruments.PutOption — Type

PutOption{T <: Widget} <: Option

abstract Option subtype. Super type for all put options types

Concrete types

Bruno.Instruments.Future — Type

Future{T <: Widget} <: FinancialInstrument

Future contract with underlying asset T.

Bruno.Instruments.EuroCallOption — Type

EuroCallOption{T <: Widget} <: CallOption{T}

European call option with underlying asset T.

Bruno.Instruments.AmericanCallOption — Type

AmericanCallOption{T <: Widget} <: CallOption{T}

American call option with underlying asset T.

Bruno.Instruments.EuroPutOption — Type

EuroPutOption{T <: Widget} <: CallOption{T}

European put option with underlying asset T.

Bruno.Instruments.AmericanPutOption — Type

AmericanPutOption{T <: Widget} <: CallOption{T}

American put option with underlying asset T.

Constructors

Base Asset Constructors

Bruno.Instruments.Stock — Method

Stock(prices, name, volatility)
Stock(;kwargs)
Stock(price; kwargs)

Construct a Stock type to use as a base asset for FinancialInstrument.

Arguments

prices:Historical prices (input as a 1-D array) or the current price input as a number <: Real

name::String: Name of the stock or stock ticker symbol. Default "".

volatility: Return volatility, measured in the standard deviation of continuous returns.

Defaults to using get_volatility() on the input prices array. Note: if a single number is given for prices volatility must be given.

Examples

Stock([1,2,3,4,5], "Test", .05)

kwargs = Dict(:prices => [1, 2, 3, 4, 5], :name => "Test", :volatility => .05);
Stock(;kwargs...)

Stock(40; volatility=.05)

Bruno.Instruments.Commodity — Method

Commodity(prices, name, volatility)
Commodity(;kwargs)
Commodity(price; kwargs)

Construct a Commodity type to use as a base asset for FinancialInstrument.

Arguments

prices:Historical prices (input as a 1-D array) or the current price input as a number <: Real

name::String: Name of the commodity or commodity ticker symbol. Default "".

volatility: Return volatility, measured in the standard deviation of continuous returns.

Defaults to using get_volatility() on the input prices array. Note: if a single number is given for prices volatility must be given.

Examples

Commodity([1,2,3,4,5], "Test", .05)

kwargs = Dict(:prices => [1, 2, 3, 4, 5], :name => "Test", :volatility => .05);
Commodity(;kwargs...)

Commodity(40; volatility=.05)

Bruno.Instruments.Bond — Method

Bond(prices, name, time_mat, coupon_rate)
Bond(;kwargs)
Bond(price; kwargs)

Construct a Bond type to use as a base asset for FinancialInstrument.

Arguments

prices:Historical prices (input as a 1-D array) or the current price input as a number <: Real

name::String: Name of the Bond or issuing company. Default "".

time_mat: Time until the bond expires (matures) in years. Default 1.

coupon_rate: The coupon rate for the bond. Default .03.

Examples

Bond([1,2,3,4,5], "Test", .5, .05)

kwargs = Dict(:prices => [1, 2, 3, 4, 5], :name => "Test", :time_mat => .5, :coupon_rate => .05);
Bond(;kwargs...)

Bond(2; coupon_rate=.05)

Financial Instrument Constructors

Missing docstring.

Missing docstring for EuroCallOption(::Widget, ::Real). Check Documenter's build log for details.

Missing docstring.

Missing docstring for AmericanCallOption(::Widget, ::Real). Check Documenter's build log for details.

Missing docstring.

Missing docstring for EuroPutOption(::Widget, ::Real). Check Documenter's build log for details.

Missing docstring.

Missing docstring for AmericanPutOption(::Widget, ::Real). Check Documenter's build log for details.